Autoload classes in GraphQL::Tracing, fix loading tracings directly
