+++ lib/graphql/tracing/null_trace.rb
@@ -1,0 +1,9 @@
+# frozen_string_literal: true
+
+require "graphql/tracing/trace"
+
+module GraphQL
+  module Tracing
+    NullTrace = Trace.new
+  end
+end
