+++ spec/graphql/autoload_spec.rb
@@ -50,13 +50,13 @@ describe GraphQL::Autoload do
   describe "loading nested files in the repo" do
     it "can load them individually" do
       files_to_load = Dir.glob("lib/**/tracing/*.rb")
-      assert_equal 24, files_to_load.size, "It found all the expected files"
+      assert_equal 26, files_to_load.size, "It found all the expected files"
       files_to_load.each do |file|
         require_path = file.sub("lib/", "").sub(".rb", "")
         stderr_and_stdout, _status = Open3.capture2e("ruby -Ilib -e 'require \"#{require_path}\"'")
         assert_equal "", stderr_and_stdout, "It loads #{require_path.inspect} in isolation"
         stderr_and_stdout, _status = Open3.capture2e("ruby -Ilib -e 'require \"graphql\"; require \"#{require_path}\"'")
-        assert_equal "", stderr_and_stdout, "It loads #{require_path.inspect} in isolation after loading graphql"
+        assert_equal "", stderr_and_stdout, "It loads #{require_path.inspect} after loading graphql"
       end
     end
   end
